wip: add sink
