+++ print.go
@@ -94,67 +94,3 @@ func Fatalf(format string, v ...interface{}) {
 	}
 }
 
-
-const (
-	// Local is the default time zone
-	LocationLocal = "Local"
-
-	// TimeFormatDefault is The default format of time
-	TimeFormatDefault = "2006-01-02 15:04:05.0000"
-)
-
-func (l *Logger) doPrint(level int, ctx Context, format string, v ...interface{}) {
-	fields := Fields{
-		Timestamp: "",
-		Level:     "",
-		Msg:       "",
-		Func:      "",
-		File:      "",
-		Line:      0,
-	}
-
-	time.LoadLocation(LocationLocal)
-	timestamp := time.Now().Format(TimeFormatDefault)
-	fields.Timestamp = timestamp
-
-	loglevel := LogLevelMap[level]
-	fields.Level = loglevel
-
-	l.mu.Lock()
-	defer l.mu.Unlock()
-
-	pc, file, line, _ := runtime.Caller(l.CallPath)
-	funcname := runtime.FuncForPC(pc).Name()
-	fields.Func = funcname
-	fields.Line = line
-
-	file = getShortFileName(file)
-	fields.File = file
-
-	var formatString string
-	if strings.EqualFold("", format) {
-		formatString = fmt.Sprint(v...)
-	} else {
-		formatString = fmt.Sprintf(format, v...)
-	}
-	fields.Msg = formatString
-
-	msg := l.formatter.Print(fields, ctx)
-	fmt.Fprintln(l.Writer, msg)
-}
-
-func (l *Logger) println(level int, ctx Context, v ...interface{}) {
-	if l.Async {
-		go l.doPrint(level, ctx, "", v...)
-	} else {
-		l.doPrint(level, ctx, "", v...)
-	}
-}
-
-func (l *Logger) printf(level int, ctx Context, format string, v ...interface{}) {
-	if l.Async {
-		go l.doPrint(level, ctx, "", v...)
-	} else {
-		l.doPrint(level, ctx, format, v...)
-	}
-}
